backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
psycopg==3.2.4
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.3.0
